## 4.3.1.0
* Bug Fixes
  * will avoid leak of local reference in Java/JNI VSI layer's seek implementation
  * corrected issues with strings found during translation to Japanese
  * migrated the name of the directory for the icons from the incorrect mil-std-2525b to mil-std-2525c
  * will resize the text widget if needed when the bloodhound is started
  * migrated over fix for GlArrow2 so that the association arrow is not render sub-terrain when no altitude/elevation is present
  * corrected issue with FAH being clipped incorrectly in areas greater than 90 degree longitude
  * corrected issue with countries using translation for All Chat Room
* Feature Additions
  * can turn off enhanced depth perception mode
  * can modify the styling of features (preliminary capability)
  * migrated to Label Manager

+++ atak/ATAK/app/src/main/java/com/atakmap/android/importfiles/ui/ImportManagerMapOverlay.java
@@ -66,7 +66,7 @@ public class ImportManagerMapOverlay extends AbstractMapOverlay2
     private final MapView _view;
     private final Context _context;
     private boolean _firstLoad = true;
-    private Persister _serializer = new Persister();
+    private final Persister _serializer = new Persister();
     private final Map<String, RemoteResource> _resources = new HashMap<>();
 
     public ImportManagerMapOverlay(MapView view,
@@ -498,7 +498,7 @@ public class ImportManagerMapOverlay extends AbstractMapOverlay2
         File resFile = new File(resource.getLocalPath());
         if (FileSystemUtils.isFile(resFile)) {
             if (FileSystemUtils.checkExtension(resFile, "zip")
-                    && resFile.getParentFile().getName()
+                    && resFile.getParentFile() != null && resFile.getParentFile().getName()
                             .equals("datapackage")) {
                 MissionPackageFileIO.deletePackage(resFile.getAbsolutePath(),
                         _view.getRootGroup());
